Apply heavy  refactor
